+++ NEWS.md
@@ -1,6 +1,10 @@
 
 # renv (development version)
 
+* Fixed an issue where `renv::hydrate()` did not hydrate packages which
+  were also listed as dependencies within a project's `DESCRIPTION` file.
+  (#1970)
+  
 * Fixed an issue where `renv::checkout()` omitted some fields from lockfile
   records when using `actions = c("snapshot", "restore")`. (#1969)